+++ src/assets/banner.xcf
