+++ src/assets/login_bg.xcf
