+++ src/store.ts
@@ -1,0 +1,3 @@
+import { writable } from 'svelte/store';
+
+export const user = writable(null);
